+++ codegen/doc.go
@@ -1,2 +1,2 @@
-// codgen helps autogenerate cloudquery plugins configured by definition
+// Package codgen helps autogenerate tables for CloudQuery source plugins from Go structs (of relevant SDKs)
 package codegen